refactor: directory layout move (no logic change) (#20)
* refactor: move codebase to app/ package (no logic change)

* fix: update remaining import references for new app/ structure

* fix: update function name references in test patches

* fix: add legacy import shims & make scheduler mock-friendly

- Add monitor.py compatibility module for legacy imports
- Update rakuten/__init__.py with module aliases and __getattr__
- Modify scheduler.py to support dependency injection for testing
- Resolves all test import errors (73 passed, 0 failed)



* style: ruff format monitor.py & scheduler.py to satisfy CI

---------
