+++ doc/BDD.md
@@ -79,4 +79,3 @@
 \*\*とき\*\* 監視ツールが通知を送信しようとすると
 \*\*ならば\*\* 通知の送信はバックオフ（時間をおいて再試行）され、最終的に失敗した場合はエラーが記録されるべきである
 \*\*かつ\*\* Prometheusの \`monitor\_fail\_total{type="discord"}\` メトリクスが1増加するべきである
-
