+++ item_db.py
@@ -1,3 +1,3 @@
 """アイテムDBエイリアス - 後方互換性のため."""
 
-from rakuten.item_db import *  # noqa: F403
+from app.db.sqlite_repo import *  # noqa: F403
