+++ rakuten_parser.py
@@ -1,3 +1,3 @@
 """楽天パーサーエイリアス - 後方互換性のため."""
 
-from rakuten.rakuten_parser import *  # noqa: F403
+from app.parser.rakuten import *  # noqa: F403
